ai-agent/requirements.txt
fastapi
uvicorn
google-generativeai
google-genai
pydantic
dotenv
chromadb